Actualización: definición de código y desarrollo de pruebas unitarias

+++ src/src/main/java/CalculadoraCientifica.java
@@ -1,16 +1,18 @@
+import java.util.InputMismatchException;
+
 class CalculadoraCientifica {
 
     public static void main(String[] args) {
 
     }
 
-    public static String ecuacionDeRecta(double x1, double y1, double x2, double y2) {
+    public static String ecuacionDeRecta(float x1, float y1, float x2, float y2) {
         try {
             if (x1 == x2) {
                 throw new ArithmeticException("¡¡ERROR DE CÁLCULO!!: El denominador no puede ser cero");
             }
-            double pendiente = ((y2 - y1) / (x2 - x1));
-            double términoN = y1 - (pendiente * x1);
+            float pendiente = ((y2 - y1) / (x2 - x1));
+            float términoN = y1 - (pendiente * x1);
             return "y = " + pendiente + "x + " + términoN;
         } catch (ArithmeticException error) {
             return error.getMessage();
